Tentative merge

+++ app/src/main/java/fr/mericfournier/unesco/model/Lieu.kt
@@ -1,5 +1,5 @@
 package fr.mericfournier.unesco.model
 
-class Lieu(var Name:String, var Id: java.lang.Integer, var Longitude: java.lang.Double, var Latitude: java.lang.Double, var Description:String) {
+class Lieu(var Name:String, var Id: Int, var Longitude: Int, var Latitude: Int, var Description:String) {
 
 }
